services/mcp_server/pyproject.toml
[project]
name = "ssyk-mcp-server"
version = "0.1.0"
description = "MCP server for Swedish occupation classification and income statistics"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "pandas",
    "numpy",
    "scikit-learn",
    "rank_bm25",
    "httpx",
    "python-dotenv",
    "openai",
    "pyarrow",
    "fastparquet"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssyk_mcp"]
